[matchingImageCollection] moved refineHomography
incidentally, hide public dependency from ceres
